+++ src/main/java/com/notenoughmail/kubejs_tfc/util/helpers/ducks/extensions/ITFCCCMoldItemBuilderExtensions.java
@@ -1,12 +1,9 @@
 package com.notenoughmail.kubejs_tfc.util.helpers.ducks.extensions;
 
 import com.notenoughmail.kubejs_tfc.item.MoldItemBuilder;
-import dev.latvian.mods.kubejs.typings.Generics;
 import dev.latvian.mods.kubejs.typings.Info;
 import dev.latvian.mods.kubejs.typings.Param;
 import dev.latvian.mods.rhino.util.RemapForJS;
-
-import java.util.List;
 
 @SuppressWarnings("unused")
 public interface ITFCCCMoldItemBuilderExtensions {
@@ -18,7 +15,6 @@ public interface ITFCCCMoldItemBuilderExtensions {
     @Info(value = "Allows the mold item to be placed in a TFC Casting With Channels mold table", params = {
             @Param(name = "model", value = "A list of 14 strings, each 14 chars long, creates the default model for the item when in the mold table")
     })
-    @Generics(String.class)
     @RemapForJS("tfcccAllowedInMoldTable")
-    MoldItemBuilder kubejs_tfc$TFCCCAllowedInMoldTable(List<String> model);
+    MoldItemBuilder kubejs_tfc$TFCCCAllowedInMoldTable(String[] model);
 }